some function refactoring
